requirements.txt
Flask==1.0
gunicorn==20.0.4
matplotlib==3.3.4
mplfinance==0.12.7a5
pandas==1.2.2
stdiomask==0.0.6
numpy==1.20.1
scipy==1.6.0
Twisted==20.3.0
APScheduler==3.7.0
python-binance==0.7.9
urllib3==1.26.3
requests==2.25.1